trunk: corrected integer kind

+++ src/base/named_array.F90
@@ -359,6 +359,7 @@ contains
 
       class(named_array3d),      intent(inout) :: this
       integer, dimension(ndims), intent(in)    :: v1, v2
+
       real,    dimension(:,:,:), pointer       :: p3d
 
       if (.not.associated(this%arr)) then
@@ -377,8 +378,9 @@ contains
 
       implicit none
 
-      class(named_array3d),            intent(inout) :: this
-      integer, dimension(ndims,LO:HI), intent(in)    :: v
+      class(named_array3d),                    intent(inout) :: this
+      integer(kind=4), dimension(ndims,LO:HI), intent(in)    :: v
+
       real,    dimension(:,:,:), pointer             :: p3d
 
       if (.not.associated(this%arr)) then
@@ -400,6 +402,7 @@ contains
       class(named_array4d),      intent(inout) :: this
       integer(kind=4),           intent(in)    :: nn
       integer, dimension(ndims), intent(in)    :: v1, v2
+
       real,    dimension(:,:,:),  pointer      :: p3d
 
       if (.not.associated(this%arr)) then
@@ -441,6 +444,7 @@ contains
 
       class(named_array4d),      intent(inout) :: this
       integer, dimension(ndims), intent(in)    :: v1, v2
+
       real,    dimension(:,:,:,:), pointer     :: p3d
 
       if (.not.associated(this%arr)) then
@@ -460,7 +464,8 @@ contains
       implicit none
 
       class(named_array4d),            intent(inout) :: this
-      integer, dimension(ndims,LO:HI), intent(in)    :: v
+      integer(kind=4), dimension(ndims,LO:HI), intent(in)    :: v
+
       real,    dimension(:,:,:,:), pointer           :: p3d
 
       if (.not.associated(this%arr)) then
